Sample illustrating how to identify direct references
This sample creates a target that's run in the project to identify
direct references and write those to a file.  This file is added to
the AdditionalFiles item, which is passed to the compiler and accessible
by analyzers.  Such an approach could be used to help an analyzer decide
which references to raise warnings about.